Correct dataset for analysis
This dataset has the weighted average of each interaction at the farm scale, the previous dataset was separated into the different habitats across the farm. There is no change to the results for the species richness based analyses (those provided in the paper), however, there are some minor changes to the plants in the optimal mixes produced when abundance and evenness are used as the optimising function (i.e., Table S1 in the associated publication).